+++ CHANGELOG.md
@@ -27,6 +27,8 @@ Compatibility:
 * Add `rb_gc_mark_locations()` (#3704, @andrykonchin).
 * Implement `rb_str_format()` (#3716, @andrykonchin).
 * Add `IO#{pread, pwrite}` methods (#3718, @andrykonchin).
+* Add `rb_io_closed_p()` (#3681, @andrykonchin).
+* Add `rb_io_open_descriptor()` (#3681, @andrykonchin).
 
 Performance: